renamed Point to FixedPoint
Having Point and DerivedPoint creates a lot of hassle because there are a lot
of situation where they should be treated the same. I have already thought
multiple times to rearrange the class hierarchy such that they inherit from
the same base class. I have tried to work around this by introducing the
DrawPoint Role, but it _was_ only a workaround. I have now decided to do it
properly.