method find_line
Started the development of a find_line method as some kind of show case
example. For the identification of points lying on an object it would actually
be more convenient have the points structure as a hash instead of an array. I
will probably change that, but I am not sure, yet, how transparent I will make
it.