broken intermediate buffering commit
I just got the idea that I could simplify the buffering approach
considerably by just clearing the buffer whenever a change is
made. This would still have the main advantage of buffering during
the evaluation of a part or the whole object tree, but would not
require the check before each read access. The only scenario where
this would be suboptimal would be a series of interleaved reads and
writes.

I commit these changes to record the step where I were: implemented
the framework in Construction and started testing. Now I will
rescind large parts of what I have done (which is not too much to
start with :-) ).