on coercion of positions into points
Started to use coercion also for the creation of Point objects from
positions when necessary. Hit a serious problem, however, which
consists of the fact that we cannot simply construct a Point
object. Instead, we need to call add_point on the construction
object, and during the coercion, the construction object is not
available.

I thought about setting the construction attribute on the point
later on. I didn't like the idea from the start, but it also is very
unpractical since add_point does not only do that. I have to think
if there is a solution to that. Otherwise I'll have to continue to
do this by hand.