testing circle line intersection with 'x'
It looks like 'x' enforces scalar context. This means that for
circle line intersection, 2 is returned instead of the two
DerivedPoint objects. This questions the whole approach, at least
with using 'x' as the intersection operator. I have to look at which
I could actually use to return a list.