backward and forward extent
The extend attribute in Role/PointSet has been changed from a Num to
a Extension (ArrayRef[Num]), holding the backward and forward
extension. I need that e.g. for rays.

When setting the attribute, this is transparent, since x is coerced
into [x, x]. When reading, the array reference itself is returned
instead of a single value.