working on Point coercion
I have to think about how to avoid that the center Point of a circle
gets autogenerated twice; maybe via a builder method for the
automatic support point in case of a fixed radius. Some how the
support point must be generated after the center has been coerced
into a Point.